+++ hyperdbg/libhyperdbg/code/debugger/core/interpreter.cpp
@@ -1102,9 +1102,10 @@ HyperDbgShowSignature()
         //
         // Debugging a special process
         //
-        ShowMessages("%x:%x u%sHyperDbg> ",
+        ShowMessages("%x:%x %s u%sHyperDbg> ",
                      g_ActiveProcessDebuggingState.ProcessId,
                      g_ActiveProcessDebuggingState.ThreadId,
+                     g_ActiveProcessDebuggingState.IsPaused ? "(paused)" : "(running)",
                      g_ActiveProcessDebuggingState.Is32Bit ? "86" : "64");
     }
     else if (g_IsSerialConnectedToRemoteDebuggee)